# Add vectorized bfloat16 atomicAdd