--- a/paddle/fluid/platform/device/gpu/gpu_primitives.h
+++ b/paddle/fluid/platform/device/gpu/gpu_primitives.h
@@ -168,12 +168,14 @@ struct VecAtomicAddHelper<platform::float16>
 #endif
 
 #ifdef PADDLE_CUDA_BF16
+#if defined(__CUDA_ARCH__) && __CUDA_ARCH__ >= 800
 template <>
 struct VecAtomicAddHelper<platform::bfloat16>
     : VecAtomicAddHelperBase<platform::bfloat16,
                              true,
                              __nv_bfloat16,
                              __nv_bfloat162> {};
+#endif
 #endif
 
 // The performance of "atomicAdd(half* )" is bad, but for "atomicAdd(half2* )"