--- a/paddle/fluid/platform/device/gpu/gpu_primitives.h
+++ b/paddle/fluid/platform/device/gpu/gpu_primitives.h
@@ -161,21 +161,19 @@ struct VecAtomicAddHelperBase {
 template <typename T>
 struct VecAtomicAddHelper : VecAtomicAddHelperBase<T, false, void, void> {};
 
-#ifdef PADDLE_CUDA_FP16
+#if CUDA_VERSION >= 10000 && defined(__CUDA_ARCH__) && __CUDA_ARCH__ >= 700
 template <>
 struct VecAtomicAddHelper<platform::float16>
     : VecAtomicAddHelperBase<platform::float16, true, __half, __half2> {};
 #endif
 
-#ifdef PADDLE_CUDA_BF16
-#if defined(__CUDA_ARCH__) && __CUDA_ARCH__ >= 800
+#if CUDA_VERSION >= 11000 && defined(__CUDA_ARCH__) && __CUDA_ARCH__ >= 800
 template <>
 struct VecAtomicAddHelper<platform::bfloat16>
     : VecAtomicAddHelperBase<platform::bfloat16,
                              true,
                              __nv_bfloat16,
                              __nv_bfloat162> {};
-#endif
 #endif
 
 // The performance of "atomicAdd(half* )" is bad, but for "atomicAdd(half2* )"