--- a/paddle/phi/backends/gpu/gpu_primitives.h
+++ b/paddle/phi/backends/gpu/gpu_primitives.h
@@ -170,7 +170,7 @@ struct VecAtomicAddHelper : VecAtomicAddHelperBase<T, false, void, void> {};
 template <>
 struct VecAtomicAddHelper<float16>
     : VecAtomicAddHelperBase<float16, true, __half, __half2> {
-  __device__ static __half zero() { return __int2half_rz(0); }
+  __device__ static __half zero() { return __half(0); }
 };
 #endif
 
@@ -178,7 +178,7 @@ struct VecAtomicAddHelper<float16>
 template <>
 struct VecAtomicAddHelper<bfloat16>
     : VecAtomicAddHelperBase<bfloat16, true, __nv_bfloat16, __nv_bfloat162> {
-  __device__ static __nv_bfloat16 zero() { return __int2bfloat16_rz(0); }
+  __device__ static __nv_bfloat16 zero() { return __nv_bfloat16(0); }
 };
 #endif