<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
Performance optimization

### PR changes
OPs

### Describe
Add vectorized atomicAdd for bfloat16 dtype. It would be used in the backward of the embedding op.

Embedding op performance testing (Backward time): id.shape = [8, 1024], weight.shape = [50304, 1024]

| | Average time (ns) | Improvement (w.r.t. Baseline FP32) |
|---|---|---|
| Baseline FP32 | 85543.9 | -- |
| Baseline BF16 | 132229.9 | -54.58% |
| BF16 (after this PR) | 50903.2 | +40.49% |